src/test/java/io/github/martinschneider/orzo/parser/ConstructorCallParserTest.java
package io.github.martinschneider.orzo.parser;

import static io.github.martinschneider.orzo.TestHelper.args;
import static io.github.martinschneider.orzo.TestHelper.assertTokenIdx;
import static io.github.martinschneider.orzo.util.Factory.*;
import static java.util.Collections.emptyList;
import static org.junit.jupiter.api.Assertions.assertEquals;

import io.github.martinschneider.orzo.error.CompilerErrors;
import io.github.martinschneider.orzo.lexer.Lexer;
import io.github.martinschneider.orzo.lexer.TokenList;
import io.github.martinschneider.orzo.parser.productions.ConstructorCall;
import java.io.IOException;
import java.util.stream.Stream;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

public class ConstructorCallParserTest {
  private ConstructorCallParser target =
      new ConstructorCallParser(ParserContext.build(new CompilerErrors()));

  private static Stream<Arguments> test() throws IOException {
    return stream(
        args("", null),
        args("new", null),
        args("new M", null),
        args("new M(", null),
        args("new M()", constr("M", emptyList())),
        args("new M(a)", constr("M", list(expr("a")))),
        args("new M(a,b,c)", constr("M", list(expr("a"), expr("b"), expr("c")))));
  }

  @ParameterizedTest
  @MethodSource
  public void test(String input, ConstructorCall expected) throws IOException {
    TokenList tokens = new Lexer().getTokens(input);
    assertEquals(expected, target.parse(tokens));
    assertTokenIdx(tokens, (expected == null));
  }
}
